requirements.txt
openai>=1.40.0
websockets>=12.0
python-dotenv>=1.0.1
pyttsx3>=2.90
